build: Add 'tools' and 'cleantools' Makefile targets.
Having a specific 'tools' target in the root Makefile allows all
the tools to be built first before the kernel or userland.

The 'cleantools' target will clean out the tools directory, and
is also called through the 'cleanall' target.

Add 'tools' to the .PHONY target.

Get rid of KCONFIG variable and target and FSUTIL target.
Point the FSUTIL variable to the fsutil binary in tools/bin.

Builds with both BSD make and GNU make.