build: Use ${INSTALL} for install(1) throughout the tree.
Using a variable for the install program allows for redefining
the variable on the make command line or in a Makefile.
This enables using other install utilities, such as NetBSD's
install(1) that has the "-U" unprivileged option.

For example, unprivileged builds are now possible with:
  $ make INSTALL="/some/path/to/NetBSD/install -U"

No specific install utility is yet defined. Whatever the make
command line or host's build system defines ${INSTALL} as is used.

Do not define ${INSTALL} in cron's Makefile.

No need to "install" groups and whoami locally during the build,
so just install them normally with the make install target.

Some variables formatting cleanup while here: $(...) to ${...}.