tools/binstall: Import OpenBSD install(1) for tools.
Import install(1) from OpenBSD 7.4 and name it binstall.

Hook into the /tools build, but do not use it yet except
for installing itself to /tools/bin.

It is time to have a local /tools version of install(1) and
not rely on the build host to supply it.

Using the local install(1) will enable unprivileged builds
once the "-U" option is implemented.

NetBSD's version of install(1) was not chosen, as it has
too many dependencies to port and features to support.
METALOG support is also not currently needed, as the fsutil
utility actually handles file and directory permissions
while creating a new file system image (sdcard.img).

OpenBSD's version is closer to 2.11BSD and 4.4BSD install(1),
the "-U" option will be trivial to implement, and it is possible
to use this version of install(1) in the DiscoBSD base system.

The idea of a local /tools/binstall is inspired from NetBSD.