tools/binstall: Implement -U unprivileged mode.
The dounpriv option is modeled after the same feature in NetBSD.
If the -U option is used, do not attempt to modify the default
owner, group, and file flags of the target.

Still do a chmod of the lower 0777 bits if unprivileged.
Copy and mask 0777 mode to tmpmode, which is used by fchmod().

Cleaned up usage() while adding -U to it.

Document the -U option in the install(1) manual page and clarify
that it is non-standard.