tools/binstall: Add conditionals to enable Linux host.
Since Linux does not have chflags(2) or issetugid(2) syscalls,
conditionally exclude calls to them with the HAVE_CHFLAGS and
HAVE_ISSETUGID defines on Linux hosts.

If not already defined, define UID_MAX and GID_MAX as UINT_MAX.

Convert from errc(3) to errx(3), and from warnc(3) to warnx(3).

Accept but silently ignore the "-f" option if not HAVE_CHFLAGS.

path_strip is _PATH_STRIP (/usr/bin/strip) if not HAVE_ISSETUGID,
and the STRIP environment variable will not be used either.

These changes enable building and running install(1) on Linux.