DiscoBSD: Update Copyright year to include v1.0-submission.
The first release of DiscoBSD was December 31, 2020, under the tag
v1.0-submission at commit cab71df177995e52106fdf1d94652bbbd6bfc6f5.
This release was the culmination of the Directed Study in 2020,
and was initially an internal release. It has since been made public.

For clarity's sake, v1.0-submission will be called the zeroth release,
as the first official release was DiscoBSD 2.0 on December 31, 2022.

Update the Copyright year for 2020.