build: Generate manual pages with DiscoBSD as OS name.
Set the operating system name in manual pages generated for DiscoBSD
by overriding the default host OS name with the mandoc -Ios option.

The ${MANROFF} variable in the system Makefile is now set to
  mandoc -Tascii -Ios="DiscoBSD ${OSREV}"
and sets the mdoc .Os macro or man .TH macro if not already set in
the source file by another operating system macro, e.g., ".UC 6".