sys: Import sys/syslimits.h header from OpenBSD.
The sys/syslimits.h header is a small 4.4BSD system header with
various system defines. It will be used to reduce namespace
pollution, especially due to the wide use of sys/param.h.

Import the whole file, but it will be pared down in another commit
to only PATH_MAX for now. It is not yet hooked into the build.