usr.bin/as: strcpy(3) -> strlcpy(3) in symbol copy.
Port over from strcpy(3) to strlcpy(3) when copying symbols into
the symbol table, and error out if there is any truncation.

Note that sizeof(stab[i].n_name) is not used as the alloc() in as
is not actually a malloc(), so it cannot be reliably used.

Keep the tools/aoututils/as version in sync with this version.

Silences the OpenBSD strcpy(3) warning when compiling on that host.