tools: Enable libbsd pkg-config variables for all tools.
The make variables ${_LIBBSD_CFLAGS} and ${_LIBBSD_LIBS} used for
Linux are now available to all tools to use, not just binstall.

The libbsd-dev package is required on Linux for BSD-specific libc
functions such as strlcpy(3). The pkg-config utility is used to
pull in the libbsd-overlay cflags and libs compile variables.