build: Add build and distribution make targets.
A 'make build', 'make all', or a plain 'make' will build all the
kernels as well as populate DESTDIR with the just built objects.

The new make target 'make distribution' will, in addition to a
'make build', build the distribution and target file system in
distrib/${MACHINE}/sdcard.img.
A 'make distribution' now will do everything that previously a
plain 'make' would do. This is now in line with the other BSDs.

All phony prerequisites are now added to the special .PHONY target.

All README.md files are updated and clarified to reflect this new
make target and build behaviour. While here, added and updated
the source tree roadmap to the project's README.md file.