build: Add release make target; update README.md.
A 'make release' will take all the objects in DESTDIR, previously
built from a 'make distribution', and create release archives in
RELEASEDIR, by default in distrib/obj/releasedir.

A release is specific to each architecture, thus a 'make clean' is
necessary between the building of each architecture.

Currently there is no 'make cleanrelease' type of target, so
RELEASEDIR must be cleaned manually before each new release.

RELEASEDIR is a new make variable, and can be redefined if needed.

Added the make variable ${OSRev} defined as ${OSMAJOR}_${OSMINOR}
and it is used in creating the release archive filenames:
tar gzip: DiscoBSD_${OSRev}_${MACHINE}.tar.gz
zip:      DiscoBSD_${OSRev}_${MACHINE}.zip

Updated README.md with documentation on building a DiscoBSD release.

Note that release directories are hardcoded in etc/Makefile and must
be kept in sync with the SUBDIR variables in sys/${MACHINE}/Makefile
for each MACHINE architecture.

This commit enables releases and snapshots to be easily generated.