libc: Add getmode(3), setmode(3) declarations to unistd.h header.
getmode(3) and setmode(3) are defined in lib/libc/gen/setmode.c
but are not declared in any of the standard libc headers, so add
it to unistd.h like all other BSDs. unistd.h already includes
sys/types.h so no need to keep it in the manual page synopsis.