usr.bin/xinstall: Force using local xerr() function.
Rename the local error function from err() to xerr(), as err() is
in libc and will create a function prototype conflict.

Remove the "if __STDC__" dance and just use the standard C code.
The old varargs was not even supported in RetroBSD, and is not
supported in DiscoBSD. Should always be moving towards POSIX.

Proper declaration for xerr(const char *, ...); and bring up the
stdarg.h include to the top with the rest of them.