usr.bin/xinstall: Do nothing for zero-sized copy; fix errno uses.
Do not have to copy if zero size supplied to copy(); just return.

Use the saved serrno value from the syscall before the call to unlink(),
as errno might belong to unlink()'s error, not the syscall before it.

Cast NULL to char * in execl(). Fixes GCC missing sentinel warning.