usr.bin/nm: Fix name clash with get_arobj() in usr.bin/ar.
ar(1) exports the function get_arobj(int fd) in archive.h.
nm(1) includes the archive.h header file, but also has an almost
duplicate copy of that function with the same name, but with
a FILE *fp parameter, as get_arobj(FILE *).

Rename get_arobj() to nm_get_arobj() to fix the name clash.

Keep the tools/aoututils/nm version in sync with this version.