tools/aoututils/ar: Use host system's strmode(3).
The binaries in /tools are built and run on the host development
system, so use the strmode(3) libc function from the host, not
ar(1)'s private one. Remove the private strmode(3) function.
The strmode(3) declaration comes from the string.h header.

strmode(3) is available on Linux through the libbsd package,
and is in standard libc for all the BSDs.

Fixes a build bug where there may be a type conflict for the
mode_t type between this code and the host's library code.

No need to make same change to ar(1) in usr.bin/ar.