tools/openbsd: Update devel/mips-elf port to OpenBSD 7.6.
The Ports Collection has made some changes since this port was made,
so update the port to modern Ports usage.

Versions are still Binutils 2.27 and GCC 4.8.1.

- Remove all RCS tags.
- s/MASTER_SITES/SITES/g
- Use -std=c++14 for building port with Clang. Clang 16.0.6 on OpenBSD
  emits many warnings and errors with the default of -std=gnu++17.
- Add patch for x_spill_indirect_levels type as unsigned char, not bool.
- Updated PLIST for devel/mips-elf/gcc.
- Add a REVISION=0 for the new patch and -std in Makefile for GCC.

Port now builds on OpenBSD/amd64 7.6.
GCC has some segfaults during the build when it is compiling libgcc,
but retrying 'make' will enable the build to finish.

Built all of DiscoBSD/pic32 with this port, and ran it on VirtualMIPS.