libc: Import double-precision floating-point library 7.0.1.
This double-precision floating-point library comes as-is from the llvm
compiler-rt builtins library, from the last release before the migration
from the MIT/Univeristy of Illinois license to the Apache 2.0 license,
Version 7.0.1, released on December 21, 2018.

https://releases.llvm.org/7.0.1/compiler-rt-7.0.1.src.tar.xz

The single-precision floating-point library code will be updated to the
same release version as the double-precision, but in a separate commit.

This is a subset of the functions available from the builtins library.
Only those desired are present in this commit. More may be added.

The double-precision floating-point code is not yet tied into the
build, and is not used. The single-precision code works and is available.