libc: Enable double-precision floating-point in libc{,_aout}.
Added the double-precision floating-point files to the list of
objects to compile for libc and libc_aout C libraries.

Sorted the list of files by single- and then double-precision.

The libc_aout libc.a works well in the pic32 port, when compiling
with a modern mips-elf-gcc compiler, such as Version 12.2.0.

GCC removed the -mshort-double compile switch for MIPS, therefore a
double-precision library is needed, not just a single-precision one.
Newer GCCs for MIPS changed the default to 64-bit doubles, not
32-bit doubles, and kept the 32-bit floats and 64-bit long doubles.

The examples in /usr/share/examples/c now do not have any "Undefined"
symbols when linking. They compile, link, and run correctly.

A modern MIPS compiler awaits!