tools/openbsd: Port devel/mips-elf GCC 12.2.0, Binutils 2.40.
mips-elf-gcc 12.2.0 does not have the -mshort-double compile switch.
A float is 32-bits; a double is 64-bits; a long double is 64-bits.
The 32-bit short double is no longer available.
Code that expects a 32-bit short double must be updated to use
either a 32-bit float or a 64-bit double.

Previous patches for forcing -std=c++14 use, x_spill_indirect_levels,
GCC passing -msoft-float to assembler, .module support, and
formatting have been added in upstream, so remove those patches.

Binutils 2.40 includes the --warn-rwx-segments linker option, which
can be turned off with --no-warn-rwx-segments.

Sync with the OpenBSD 7.6 Ports version of arm-none-eabi 12.2.0/2.40,
not including the clang-on-powerpc64 patch.