tools/kconfig: Fix Copyright years typo in mkmakefile.c.
The commit a0c84ff41192d6f40def3a59b9ac52fa4289bf56 from the 4.4BSD
sources available at https://github.com/dspinellis/unix-history-repo
states "4.4BSD snapshot (revision 8.1); add 1993 to copyright"
wherein the year "1993" was to be added to the years "1980,1990".

It is clear by inspection that the years should be "1980, 1990, 1993"
and not "1993, 19801990".

Use the Copyright years "1980, 1990, 1993" to fix the typo.

No change has been made to the Copyright.