tools/fsutil: Use strlcpy(3); proper path[] lengths.
The manifest code handled the path[] members of struct entry_t and
struct link_info_t in a workable, but non-standard, way that abuses
strcpy(3)'s ability to write past the end of a destination buffer.

Converting to strlcpy(3) broke this non-standard approach by always
filling the path member's single char array with a terminating NUL.

The size of path[] members and the usage of malloc(3) must change.

Increase path[] struct members of entry_t and link_info_t from one
character arrays to BSDFS_MAXNAMLEN+1 length character arrays.

Define BSDFS_MAXNAMLEN as 63, which is the MAXNAMLEN of 2.11BSD and
all 2.11BSD-derived operating systems like RetroBSD and DiscoBSD.

malloc(3) only the size of struct entry_t or struct link_info_t.
Do not include the length of path in the supplied size parameter.

Check the return value from strlcpy(3) for truncation, and if it
happens, print an error message to stderr and continue.

No need to duplicate the function body of ftsent_compare().