tools/openbsd: GCC 12.2.0 on OpenBSD 7.6.
Update GCC from the Linaro version 7.4-2019.02 to the standard
GCC version 12.2.0 from the Ports Collection for OpenBSD 7.6.

Put this in the .../mystuff/devel/arm-none-eabi/gcc directory,
as it will be a custom port once multilib support is added.

Updates to newlib and GDB will be in a separate commit.

- Remove all RCS tags.
- Use SITE_GCC.
- Use .tar.xz archive suffix instead of .tar.bz2.
- Update patches and packing lists.

This version of the port does not have the multilib support.
It will be added back in a separate commit.