tools/openbsd: Enable arm-none-eabi-gcc rmprofile multilib.
Update to the arm thumb rmprofile multilib for GCC 12.2.0 in
.../mystuff/devel/arm-none-eabi/gcc and enable it.

Bump REVISION to 2, since OpenBSD Ports is at 1.

Enable building the multilib runtime libraries with the 'rmprofile'
option for configuring arm-none-eabi-gcc with --with-multilib-list.

This splits out separate libgcc.a runtime libraries for all the
Cortex-M and Cortex-R architecture profiles (i.e., v6-m, v7-m,
v7e-m, v7-r) with none, soft, and hard floating point support.

Hand massage PFRAG.arm packing list, as update-plist has a hard
time dealing with the FLAVORS for this port.

Not adding 'aprofile' support, as it can conflict with 'rmprofile'
libraries, and it is not required for current development needs.

DiscoBSD binaries built with this arm-none-eabi-gcc port now run
correctly and do not hard fault on undefined instructions, as was
the case without 'rmprofile' multilib support built in.