tools/openbsd: Remove old arm-none-eabi port.
Remove the old multi-port of arm-none-eabi from OpenBSD 6.6, which
included GCC linaro 7.4-2019.02, binutils 2.27, newlib 2.2.0-1 and
GDB 7.9.1 from the standard OpenBSD Ports Collection.

It was saved here for posterity. but is now not up to date and does
not include the patches for rmprofile for the runtime libraries.

It should not be used for current development. Instead, use the
custom port in tools/openbsd/ports/mystuff/devel/arm-none-eabi.