build: Consistent libbsd-overlay use on Linux hosts.
Add CFLAGS and LIBS sourced from pkg-config for libbsd-overlay on
Linux hosts building executables to run on the host, mainly tools.

The package used for the BSD C library is libbsd-dev on Ubuntu.

More consistency in supplying the CFLAGS and LIBS now that they
are added to share/mk/sys.mk, the system include makefile.

Fixes build issues on Linux hosts with the strcpy(3) -> strlcpy(3)
commits and the similar types of updates and changes.