tools/kconfig: strcpy(3) -> strlcpy(3); use libbsd.
Use the strlcpy(3) bounded string function instead of strcpy(3).

Straight update to the safer function without implementing
error-checking and handling of possible truncation.

The uses are from a fixed hardcoded string, not untrusted input.

strlcpy(3) is available on Linux through the libbsd-dev package,
and is in standard libc for all the BSDs.

Silences the OpenBSD host linker warning for strcpy(3).