sys: Use sys/sys as system header file directory.
Align with 4.4BSD to have the system/kernel header files in sys/sys,
not in sys/include.

Update the new location in the kernel Config templates for stm32
and pic32, and generate new kernel Makefiles for each target board.

Use the relative system directory variable $S in link in Makefiles.

Update new location in include/Makefile when copying include files,
in sys/conf/newvers.sh, and the sys symlink in /include directory.