share/man: Enable Section 9; add intro.9 manpage.
The Makefile for Section 9 is based on the Section 7 Makefile,
with required changes for Section 9.

Added cat9 to the lists of all manpage directories in Makefiles.

The intro.9 manual page is imported from OpenBSD verbatim,
copyright Michael Shalayeff.
It will be updated from OpenBSD to DiscoBSD in a future commit.

Added the generated intro.0 manpage to the fsutil base/mi spec.