share/man: Update intro.9 for DiscoBSD.
The ".Ox" macro specifies "OpenBSD". The ".Bx" macro specifies "BSD".

So ".Bx Disco" gives the desired result "DiscoBSD".

Note that a version number as a second argument as ".Bx Disco 2.5"
will not give the desired result of "DiscoBSD 2.5", but of the
hyphenated "DiscoBSD-2.5", as used for "4.3BSD-Tahoe".

Put the version number on the following line by itself.