share/man: Update intro.6 for DiscoBSD; link ppt(6).
The bcd(6) manual page contains the ppt(6) game utility as well,
so link ppt(6) to bcd(6) in its Makefile and the fsutil base/mi spec.

Update intro.6 for all games currently available in DiscoBSD,
and cross reference the games that have a manual page.

Remove references to games that are not yet installed nor available.

2.11BSD historically kept game data files under /usr/games/lib.

Update the HISTORY section for DiscoBSD.

Not using Mdocdate, as DiscoBSD is not under CVS version control.