pic32: Initialize d_flags field in struct bdevsw.
The d_flags field of struct bdevsw was not being initialized and
relied on the compiler to initialize it.
The d_flags field is only used for magtape, so initialize to zero
for all drivers.

Silences GCC -Wmissing-field-initializers warnings.

No binary change. Compiler was already initializing d_flags to zero.