stm32: Add MPU reading functions to LL_CORTEX library.
Add unofficial extensions to LL_MPU_* functions to read from and write
to the MPU Control register, and read from the MPU Type register.

The LL_CORTEX library does not offer an API to read MPU registers,
so they had to be implemented here independently.

These functions will be used by sysctl in a future commit, to easily
interrogate and set up the Memory Protection Unit.